Less insane templating and routing (Express+hbs)
