+++ index.js
@@ -1,278 +1,30 @@
-const path = require('path');
-const _ = require('lodash');
-const fse = require('fs-extra');
-const Hapi = require('hapi');
-const config = require('./config')
-const requestLib = require('request');
+const express = require('express');
+const config = require('./config');
+const routes = require('./routes');
+const fmt = require('./lib/fmt');
 
-const VIEWSDIR = config.paths.views;
-const JSONDIR = config.paths.json;
-const LOGFILE = config.paths.logfile;
+function startApp () {
+  const app = express();
+  app.disable('x-powered-by');
 
-let GH_CACHE;
+  // register hbs
+  const hbs = config.views.engine;
+  hbs.registerPartials(config.views.partialsPath);
+  app.engine('html', hbs.__express);
+  app.set('view engine', 'html');
 
-const server = new Hapi.Server();
-server.connection({port: config.appPort});
+  routes(app);
 
-function _currentTime () {
-  const now = new Date();
-  const date = [now.getFullYear(), now.getMonth() + 1, now.getDate()].join('-');
-  return date;
+  if (config.appNetwork === 'private' && !config.appNetworkInterface) {
+    console.log('Couldnt determine private interface - restricting to localhost');
+    config.appNetworkInterface = '127.0.0.1';
+  }
+  // start app
+  app.listen(config.appPort, config.appNetworkInterface, 0, function () {
+    console.log(`[${fmt.date()} ${fmt.time()}] Great Converto listening on port ${config.appPort}`);
+  });
+
+  return app;
 }
 
-// TODO: Don't replace dots in paths
-// TODO: Aggregate 404s
-function _logRequest(url) {
-  if (url === '/favicon.ico' || url === '/robots.txt') {
-    return;
-  }
-  const key = url.replace('.', '-') + '.' + _currentTime();
-
-  const data = fse.readJsonSync(LOGFILE) || {};
-  const currentCount = _.get(data, key) || 0;
-  _.set(data, key, currentCount + 1);
-
-  fse.writeJsonSync(LOGFILE, data);
-}
-
-// TODO: read https://api.github.com/repos/vot/ffbinaries-prebuilt/releases
-// to get stats + cache for 10 min
-function _replacePlaceholders(data, payload) {
-  data = data.replace(/%%BASEURL%%/g, config.baseUrl);
-  data = data.replace(/%%CURRENT_VERSION%%/g, config.currentVersion);
-
-  if (payload) {
-    _.each(payload, function (val, key) {
-      data = data.replace('{{'+key+'}}', val);
-    });
-  }
-
-  return data;
-}
-
-function _getJson(file) {
-  let data;
-
-  try {
-    data = fse.readFileSync(path.resolve(JSONDIR + '/' + file + '.json')).toString();
-    data = _replacePlaceholders(data);
-    data = JSON.parse(data);
-  } catch (e) {
-    console.log(e);
-    data = '404';
-  }
-  return data;
-}
-
-function _getView(file, payload) {
-  let data;
-  try {
-    data = fse.readFileSync(VIEWSDIR + '/' + file + '.html').toString();
-    data = _replacePlaceholders(data, payload);
-  } catch (e) {
-    console.log(e);
-    data = '404';
-  }
-  return data;
-}
-
-const preResponse = function (request, reply) {
-  _logRequest(request.url.path);
-  return reply.continue();
-};
-
-server.ext('onPreResponse', preResponse);
-
-// Info pages
-server.route({
-  method: 'GET',
-  path: '/',
-  handler: function (request, reply) {
-    reply(_getView('home'));
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/readme',
-  handler: function (request, reply) {
-    reply(_getView('readme'));
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/downloads',
-  handler: function (request, reply) {
-    reply(_getView('downloads'));
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/stats',
-  handler: function (request, reply) {
-    const requestLog = _getJson('../requestlog');
-    let requestMap = {};
-    _.each(requestLog, function (val, key) {
-      requestMap[key] = _.sum(_.map(val));
-    });
-    var pagePaths = [
-      "/",
-      "/stats",
-      "/readme"
-    ];
-
-    var logPageRequests = _.pick(requestMap, pagePaths);
-
-    var logApiRequests = _.pickBy(requestMap, function (v, k) {
-      return k.startsWith('/api')
-    });
-
-    function objectToHtml (object) {
-      var rtn = '<ul class="stats">';
-
-      _.each(object, function (val, k) {
-        rtn += '<li>';
-        rtn += '<span class="count">' + val + '</span> ';
-        rtn += '<code>' + k + '</code>';
-        rtn += '</li>';
-      });
-
-      rtn += '</ul>';
-      return rtn;
-    }
-
-    let payload = {
-      // logApiRequests: JSON.stringify(logApiRequests, null, 2),
-      logApiRequests: objectToHtml(logApiRequests),
-      logApiRequestsTotal: _.sum(_.map(logApiRequests)),
-
-      // logPageRequests: JSON.stringify(logPageRequests, null, 2),
-      logPageRequests: objectToHtml(logPageRequests),
-      logPageRequestsTotal: _.sum(_.map(logPageRequests)),
-    };
-
-    if (GH_CACHE && GH_CACHE.expiration > Date.now()) {
-      payload.github = objectToHtml(GH_CACHE.data);
-      payload.githubTotal = GH_CACHE.data.total;
-      return reply(_getView('stats', payload));
-    }
-
-    const requestOpts = {
-      url: 'https://api.github.com/repos/vot/ffbinaries-prebuilt/releases',
-      headers: {
-        'User-Agent': 'ffbinaries.com'
-      }
-    }
-
-    requestLib(requestOpts, function (error, response, body) {
-      if (!error && response.statusCode == 200) {
-        var ghData = JSON.parse(body);
-        var ghDataRtn = {
-          total: 0
-        };
-        _.each(ghData, function (release) {
-          _.each(release.assets, function (val) {
-            ghDataRtn[val.name] = val.download_count;
-            ghDataRtn.total += val.download_count;
-          });
-        });
-
-        // var ghDataString = JSON.stringify(ghDataRtn, null, 2);
-        GH_CACHE = {
-          expiration: Date.now() + (60 * 1000),
-          data: ghDataRtn
-        }
-
-        payload.github = objectToHtml(ghDataRtn);
-        payload.githubTotal = ghDataRtn.total;
-      }
-
-      reply(_getView('stats', payload));
-    })
-  }
-});
-
-// API V1 redirect /api requests to /api/v1
-server.route({
-  method: 'GET',
-  path: '/api',
-  handler: function (request, reply) {
-    // reply.redirect('/api/v1');
-    reply(_getView('api'));
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/api/versions',
-  handler: function (request, reply) {
-    reply.redirect('/api/v1');
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/api/latest',
-  handler: function (request, reply) {
-    reply.redirect('/api/v1/version/latest');
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/api/version/{version}',
-  handler: function (request, reply) {
-    reply.redirect('/api/v1/version/' + request.params.version);
-  }
-});
-
-
-// API V1 versions
-server.route({
-  method: 'GET',
-  path: '/api/v1',
-  handler: function (request, reply) {
-    reply(_getJson('index'));
-  }
-});
-
-// versions
-server.route({
-  method: 'GET',
-  path: '/api/v1/versions',
-  handler: function (request, reply) {
-    reply.redirect('/api/v1');
-  }
-});
-server.route({
-  method: 'GET',
-  path: '/api/v1/latest',
-  handler: function (request, reply) {
-    reply.redirect('/api/v1/version/latest');
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/api/v1/version/latest',
-  handler: function (request, reply) {
-    reply(_getJson(config.currentVersion));
-  }
-});
-
-server.route({
-  method: 'GET',
-  path: '/api/v1/version/{version}',
-  handler: function (request, reply) {
-    reply(_getJson(request.params.version));
-  }
-});
-
-server.start((err) => {
-  if (err) {
-    throw err;
-  }
-  console.log(`Server running at: ${server.info.uri}`);
-});
+startApp();
