+++ views/downloads.html
@@ -1,29 +1,3 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-  <title>Download binaries for ffmpeg, ffprobe, ffserver and ffplay (cross-platform: Windows, Mac, Linux)</title>
-  <link href='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css' rel='stylesheet' type='text/css'>
-</head>
-<body>
-
-  <nav class="navbar navbar-default navbar-inverse navbar-fixed-top">
-    <div class="container">
-      <ul class="nav navbar-nav">
-        <li><a class="navbar-brand" href="/">ffbinaries</a></li>
-        <li><a href="/downloads">Downloads</a></li>
-        <li><a href="/api">API</a></li>
-        <li><a href="/readme">Readme</a></li>
-        <li><a href="/stats">Stats</a></li>
-        <li><a href="https://www.npmjs.com/package/ffbinaries" target="_blank">Node.js module</a></li>
-      </ul>
-    </div>
-  </nav>
-
-  <br /><br /><br />
-
-  <div class="container">
     <h1>Download binary files for ffmpeg, ffprobe, ffserver and ffplay</h1>
 
     <br />
@@ -152,15 +126,3 @@
       </tr>
 
     </table>
-
-  </div>
-
-  <style>
-    @media (max-width: 767px) {
-      .navbar {position:relative}
-      .navbar-brand {float: none}
-    }
-  </style>
-
-</body>
-</html>
