+++ views/layout.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html>
+{{>head}}
+<body>
+
+  <nav class="navbar navbar-default navbar-inverse navbar-fixed-top">
+    <div class="container">
+      <ul class="nav navbar-nav">
+        <li><a class="navbar-brand" href="/">ffbinaries</a></li>
+        <li><a href="/downloads">Downloads</a></li>
+        <li><a href="/api">API</a></li>
+        <li><a href="/readme">Readme</a></li>
+        <li><a href="/stats">Stats</a></li>
+        <li><a href="https://www.npmjs.com/package/ffbinaries" target="_blank">Node.js module</a></li>
+      </ul>
+    </div>
+  </nav>
+
+  <br /><br /><br />
+
+  <div class="container">
+    {{{body}}}
+
+  </div>
+
+  <style>
+    ul.stats {padding: 0; margin: 0; list-style: none}
+    ul.stats li {line-height: 26px; padding: 0; margin: 0; list-style: none}
+    ul.stats .count {width: 50px; text-align: right; font-size: 15px; font-weight: bold; display: inline-block}
+
+    @media (max-width: 767px) {
+      .navbar {position:relative}
+      .navbar-brand {float: none}
+    }
+  </style>
+
+</body>
+</html>
