+++ views/partials/head.html
@@ -1,0 +1,8 @@
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+  <title>Download binaries for ffmpeg, ffprobe, ffserver and ffplay (cross-platform: Windows, Mac, Linux)</title>
+  <link href='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css' rel='stylesheet' type='text/css'>
+  <!-- <link href="/public/css/bootstrap.min.css" rel="stylesheet" type="text/css"> -->
+  <!-- <link href="/public/css/ffbinaries.css" rel="stylesheet" type="text/css"> -->
+</head>
